Abstract access to the tracker via a per-object recorder.
The recorder manages the path on behalf of the tracked object and
provides a nice API (create(), edit(), remove()) for the tracked object.

Critically, relieving the tracked object of managing the path allows the
path to be changed in a central location, e.g. when an item is removed
from a list the following items need their paths "fixed".