Track the previous or removed value ('was') so the actual change is obvious.
For instance a change to a string in a dict, from 'foo' to 'bar', looks
like:

  {'action': 'edit',
   'path': [..., 'string'],
   'value': 'bar',
   'was': 'foo'}

Unfortunately, the 'was' value must be the value when the document was
loaded, i.e. any changes made to the value up to the point it's removed
must be undone.

Now, we do have the steps to undo the change ... but ...  yikes!