Move the track() generic func to the Tracker instance to avoid unnecessary Recorder creation.
Immutable objects don't need a Recorder instance but the old track()
functions required a Recorder instance as an arg. Leaving it to the
Tracker instance means it can create a Recorder only if needed, i.e. not
for immutable objects.

The Recorded has also become starting to track child objects, although
it was basically doing that already but more along the lines of "create
me a recorder I can use" instead of "track this, please". Nicer, I
think.