Call a hook function, passing all the changes, to a hook function before actually writing to the database.
The hook function may make additional changes (that's what it's for in
fact) and so the state of the session needs to be "frozen" before the
hook is called and then checked for additional changes.

The total set of updates to write to CouchDB is all of the changes made
during the session and during all the pre-flush hook calls.