Extend the sesson to have pre and post flush hooks.
The signature of the flush hooks is the same, it's just when the hook
gets called that changes.

A pre flush hook is useful when everything you need to do is in memory,
i.e. you don't need to look at the content of the CouchDB. A post flush
hook is used when the state of the CouchDB affects the processing, i.e.
if you rely on a view.