Fix the flush hook deletion generator.
The deletion generator was trying to find the document in the cache,
only that's not where deleted documents appear. No tests triggered the
problem because there were no tests that consumed all the flush
generators.

Now, the deleted document is stored in the _deleted dict instead of just
the document's rev.