Remove the deepcopy from session create
couchdb-python takes copies of documents before creating. This would mean
having to add a copy method to file objects. We figure breaking with
couchdb-pythons api is better than making a copy of files? We also
discussed that it's easy to make a copy in your own code if you want to
isolate your data.